Refactor rdfframework as multi-module Python package
