+++ rdfframework/utilities/debug.py
