+++ tests/__init__.py
@@ -1,5 +1,6 @@
 import unittest
-from test_rdfframework import *
+#from test_rdfframework import *
+from test_validators import *
 
 if __name__ == '__main__':
     unittest.main()
